+++ quantify_scheduler/backends/qblox/compiler_abc.py
@@ -1637,7 +1637,7 @@ class QbloxBaseModule(ControlDeviceCompiler, ABC):
                     self._settings.in1_gain = in1_gain
                 else:
                     raise ValueError(
-                        f"Overwriting gain of {io_name} of module {self.name}"
+                        f"Overwriting gain of {io_name} of module {self.name} "
                         f"to in1_gain: {in1_gain}.\nIt was previously set to "
                         f"in1_gain: {self._settings.in1_gain}."
                     )